esl-iter: simplify buffer pointer arithmetic
The ((intptr_t)iter->buf + (unsigned long)iter->offset) cast dance in
esl_list_iter_next_with_size_correction() is equivalent to plain
pointer arithmetic on iter->buf with (size_t)iter->offset, and avoids
the unsigned long cast which isn't portable for off_t on 32-bit.

+++ src/esl-iter.c
@@ -395,8 +395,7 @@ esl_list_iter_next_with_size_correction(esl_list_iter *iter, efi_guid_t *type,
 		iter->offset += iter->esl->signature_list_size;
 		if ((uint32_t)iter->offset >= iter->len)
 			return 0;
-		iter->esl = (efi_signature_list_t *)((intptr_t)iter->buf
-						+ (unsigned long)iter->offset);
+		iter->esl = (efi_signature_list_t *)(iter->buf + (size_t)iter->offset);
 	}
 
 	efi_signature_list_t esl;
